+++ packages/cluster/src/commands/down.js
@@ -4,33 +4,39 @@ const { exec, reporter } = require('@dhis2/cli-helpers-engine')
 const {
     initDockerComposeCache,
     makeComposeProject,
-    makeDockerImage,
+    resolveConfiguration,
 } = require('../common')
 
-const run = async function({ v, clean, getCache, ...argv }) {
+const defaults = require('../defaults')
+
+const run = async function({ name, clean, getCache, cluster, ...argv }) {
+    const { dockerComposeRepository } = resolveConfiguration(argv, {}, cluster)
     const cacheLocation = await initDockerComposeCache({
         cache: getCache(),
-        dockerComposeRepository: argv.cluster.dockerComposeRepository,
+        dockerComposeRepository,
         force: false,
     })
+
     if (!cacheLocation) {
         reporter.error('Failed to initialize cache...')
         process.exit(1)
     }
-    console.log(`Winding down cluster version ${chalk.cyan(v)}`)
+
+    console.log(`Winding down cluster ${chalk.cyan(name)}`)
     try {
         await exec({
             cmd: 'docker-compose',
             args: [
                 '-p',
-                makeComposeProject(v),
+                makeComposeProject(name),
                 '-f',
                 path.join(cacheLocation, 'docker-compose.yml'),
                 'down',
             ].concat(clean ? ['--volumes'] : []),
             env: {
-                DHIS2_CORE_TAG: makeDockerImage(v),
-                DHIS2_CORE_PORT: '0000', // Doesn't matter for `down`
+                DHIS2_CORE_IMAGE: defaults.image,
+                DHIS2_CORE_NAME: name,
+                DHIS2_CORE_PORT: defaults.port,
             },
         })
     } catch (e) {
@@ -40,16 +46,10 @@ const run = async function({ v, clean, getCache, ...argv }) {
 }
 
 module.exports = {
-    command: 'down <v>',
+    command: 'down <name>',
     desc: 'Destroy a running container',
     aliases: 'd',
     builder: {
-        port: {
-            alias: 'p',
-            desc: 'Specify the port on which to expose the DHIS2 instance',
-            type: 'integer',
-            default: 8080,
-        },
         clean: {
             desc: 'Destroy all data volumes as well as ephemeral containers',
             type: 'boolean',