+++ packages/cluster/src/commands/restart.js
@@ -4,33 +4,36 @@ const { reporter, exec, tryCatchAsync } = require('@dhis2/cli-helpers-engine')
 const {
     initDockerComposeCache,
     makeComposeProject,
-    makeDockerImage,
+    resolveConfiguration,
 } = require('../common')
 
-const run = async function({ service, v, port, ...argv }) {
+const defaults = require('../defaults')
+
+const run = async function({ service, name, port, cluster, ...argv }) {
+    const { dockerComposeRepository } = resolveConfiguration(argv, {}, cluster)
     const cacheLocation = await initDockerComposeCache({
         cache: argv.getCache(),
-        dockerComposeRepository: argv.cluster.dockerComposeRepository,
+        dockerComposeRepository,
         force: false,
     })
     if (!cacheLocation) {
         reporter.error('Failed to initialize cache...')
         process.exit(1)
     }
-    reporter.info(`Spinning up cluster version ${chalk.cyan(v)}`)
+    reporter.info(`Spinning up cluster version ${chalk.cyan(name)}`)
     const res = await tryCatchAsync(
         'exec(docker-compose)',
         exec({
             cmd: 'docker-compose',
             args: [
                 '-p',
-                makeComposeProject(v),
+                makeComposeProject(name),
                 '-f',
                 path.join(cacheLocation, 'docker-compose.yml'),
                 'restart',
             ].concat(service ? [service] : []),
             env: {
-                DHIS2_CORE_TAG: makeDockerImage(v),
+                DHIS2_CORE_NAME: name,
                 DHIS2_CORE_PORT: port,
             },
             pipe: true,
@@ -43,15 +46,16 @@ const run = async function({ service, v, port, ...argv }) {
 }
 
 module.exports = {
-    command: 'restart <v> [service]',
+    command: 'restart <name> [service]',
     desc: 'Restart a cluster or cluster service',
     aliases: 'r',
     builder: {
         port: {
             alias: 'p',
-            desc: 'Specify the port on which to expose the DHIS2 instance',
+            desc: `Specify the port on which to expose the DHIS2 instance (default: ${
+                defaults.port
+            })`,
             type: 'integer',
-            default: 8080,
         },
     },
     handler: run,