+++ packages/cluster/src/commands/seed.js
@@ -1,13 +1,23 @@
 const { reporter } = require('@dhis2/cli-helpers-engine')
-const { initDockerComposeCache } = require('../common')
+const { initDockerComposeCache, resolveConfiguration } = require('../common')
 const { seed } = require('../db')
 
-const run = async function(argv) {
+const defaults = require('../defaults')
+
+const run = async function({ dhis2Version, ...argv }) {
+    const { cluster } = argv
+    const { dockerComposeRepository, dbVersion } = resolveConfiguration(
+        argv,
+        {},
+        cluster
+    )
+
     const cacheLocation = await initDockerComposeCache({
         cache: argv.getCache(),
-        dockerComposeRepository: argv.cluster.dockerComposeRepository,
+        dockerComposeRepository,
         force: false,
     })
+
     if (!cacheLocation) {
         reporter.error('Failed to initialize cache...')
         process.exit(1)
@@ -15,12 +25,13 @@ const run = async function(argv) {
 
     return await seed({
         cacheLocation,
+        dbVersion,
         ...argv,
     })
 }
 
 module.exports = {
-    command: 'seed <v> [path]',
+    command: 'seed <name> [path]',
     desc: 'Seed the database from a backup',
     builder: {
         update: {
@@ -28,6 +39,10 @@ module.exports = {
             desc: 'Force re-download of cached files',
             type: 'boolean',
             default: false,
+        },
+        dhis2Version: {
+            desc: 'DHIS2 version to use',
+            type: 'string',
         },
     },
     handler: run,