+++ packages/cluster/src/commands/up.js
@@ -4,14 +4,20 @@ const { reporter, exec, tryCatchAsync } = require('@dhis2/cli-helpers-engine')
 const {
     initDockerComposeCache,
     makeComposeProject,
-    makeDockerImage,
+    makeEnvironment,
+    resolveConfiguration,
 } = require('../common')
+
+const defaults = require('../defaults')
 const { seed: doSeed } = require('../db')
 
-const run = async function({ v, port, seed, seedFile, update, ...argv }) {
+const run = async function(argv) {
+    const { cluster, name, seed, seedFile, update } = argv
+    const cfg = resolveConfiguration(argv, {}, cluster)
+
     const cacheLocation = await initDockerComposeCache({
         cache: argv.getCache(),
-        dockerComposeRepository: argv.cluster.dockerComposeRepository,
+        dockerComposeRepository: cfg.dockerComposeRepository,
         force: update,
     })
 
@@ -21,27 +27,32 @@ const run = async function({ v, port, seed, seedFile, update, ...argv }) {
     }
 
     if (seed || seedFile) {
-        await doSeed({ cacheLocation, v, path: seedFile, update, ...argv })
+        await doSeed({
+            name,
+            cacheLocation,
+            dbVersion: cfg.dbVersion,
+            url: cfg.demoDatabaseURL,
+            path: seedFile,
+            update,
+            ...argv,
+        })
     }
 
-    reporter.info(`Spinning up cluster version ${chalk.cyan(v)}`)
+    reporter.info(`Spinning up cluster ${chalk.cyan(name)}`)
+
     const res = await tryCatchAsync(
         'exec(docker-compose)',
         exec({
+            env: makeEnvironment(cfg),
             cmd: 'docker-compose',
             args: [
                 '-p',
-                makeComposeProject(v),
+                makeComposeProject(name),
                 '-f',
                 path.join(cacheLocation, 'docker-compose.yml'),
                 'up',
                 '-d',
             ],
-            env: {
-                DHIS2_CORE_TAG: makeDockerImage(v),
-                DHIS2_CORE_VERSION: v,
-                DHIS2_CORE_PORT: port,
-            },
             pipe: true,
         })
     )
@@ -52,21 +63,21 @@ const run = async function({ v, port, seed, seedFile, update, ...argv }) {
 }
 
 module.exports = {
-    command: 'up <v>',
+    command: 'up <name>',
     desc: 'Spin up a new cluster',
     aliases: 'u',
     builder: {
         port: {
             alias: 'p',
-            desc: 'Specify the port on which to expose the DHIS2 instance',
+            desc: `Specify the port on which to expose the DHIS2 instance (default: ${
+                defaults.port
+            })`,
             type: 'integer',
-            default: 8080,
         },
         seed: {
             alias: 's',
             desc: 'Seed the detabase from a sql dump',
             type: 'boolean',
-            default: false,
         },
         seedFile: {
             desc:
@@ -77,7 +88,34 @@ module.exports = {
             alias: 'u',
             desc: 'Indicate that d2 cluster should re-download cached files',
             type: 'boolean',
-            default: false,
+        },
+        image: {
+            alias: 'i',
+            desc: 'Specify the Docker image to use',
+            type: 'string',
+        },
+        dhis2Version: {
+            desc: 'Set the DHIS2 version',
+            type: 'string',
+        },
+        dbVersion: {
+            desc: 'Set the database version',
+            type: 'string',
+        },
+        channel: {
+            desc: `Set the release channel to use (default: ${
+                defaults.channel
+            })`,
+            type: 'string',
+        },
+        customContext: {
+            alias: 'c',
+            desc: 'Serve on a custom context path',
+            type: 'boolean',
+        },
+        variant: {
+            desc: 'Append variant options to the image',
+            type: 'string',
         },
     },
     handler: run,